refactor: base entity 통일
